+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,10 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=/Users/victor/Documents/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/victor/Documents/GitHub/share_social_media_plugin/example"
-export "FLUTTER_TARGET=/Users/victor/Documents/GitHub/share_social_media_plugin/example/lib/main.dart"
+export "FLUTTER_APPLICATION_PATH=/Users/victor/Documents/ITERA/share_social_media_plugin/example"
+export "FLUTTER_TARGET=/Users/victor/Documents/ITERA/share_social_media_plugin/example/lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+export "OTHER_LDFLAGS=$(inherited) -framework Flutter"
 export "FLUTTER_FRAMEWORK_DIR=/Users/victor/Documents/flutter/bin/cache/artifacts/engine/ios"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"